Crate actix_hash

Hashing utilities for Actix Web.

Crate Features

All features are enabled by default.

blake2: Blake2 types

md4: MD4 types 🚩

md5: MD5 types 🚩

sha1: SHA-1 types 🚩

sha2: SHA-2 types

sha3: SHA-3 types

Security Warning 🚩

The md4, md5, and sha1 types are included for completeness and interoperability but they are considered cryptographically broken by modern standards. For security critical use cases, you should move to using the other algorithms.

Structs

BodyHash

Wraps an extractor and calculates a body checksum hash alongside.

BodyHashParts

Parts of the resulting body hash extractor.

Type Definitions

BodyBlake2b

Wraps an extractor and calculates a Blake2b body checksum hash alongside.

BodyBlake2s

Wraps an extractor and calculates a Blake2s body checksum hash alongside.

BodyMd4

Wraps an extractor and calculates a MD4 body checksum hash alongside.

BodyMd5

Wraps an extractor and calculates a MD5 body checksum hash alongside.

BodySha1

Wraps an extractor and calculates a SHA-1 body checksum hash alongside.

BodySha3_224

Wraps an extractor and calculates a SHA-3-224 body checksum hash alongside.

BodySha3_256

Wraps an extractor and calculates a SHA-3-256 body checksum hash alongside.

BodySha3_384

Wraps an extractor and calculates a SHA-3-384 body checksum hash alongside.

BodySha3_512

Wraps an extractor and calculates a SHA-3-512 body checksum hash alongside.

BodySha224

Wraps an extractor and calculates a SHA-224 body checksum hash alongside.

BodySha256

Wraps an extractor and calculates a SHA-256 body checksum hash alongside.

BodySha384

Wraps an extractor and calculates a SHA-384 body checksum hash alongside.

BodySha512

Wraps an extractor and calculates a SHA-512 body checksum hash alongside.